+++ NEWS.md
@@ -1,10 +1,12 @@
 TimeStructures release notes
 ===================================
 
-Version 0.1.8 (2021-08-02)
+Version 0.2.0 (2021-08-03)
 --------------------------
 * Restructured and commented the file for improved readability (#7)
 * Inclusion of functions for DynamicTwoLevel and DynamicOperationalLevel (#7)
+* Inclusion of DynamicStrategicLevel (#7)
+* Export of all two level time structures (#7)
 
 Version 0.1.7 (2021-06-14)
 --------------------------